% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-eda.R
\name{plot_horizontal_dotplot}
\alias{plot_horizontal_dotplot}
\title{Plot horizontal dot plot.}
\usage{
plot_horizontal_dotplot(
  data,
  x_str,
  y_str,
  color_str = NULL,
  line_size = 1,
  line_color = "darkgray",
  point_size = 8,
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string (optional). Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{line_size}{Size of line, representing range of values.}

\item{line_color}{Color of line, representing range of values.}

\item{point_size}{Size of point.}

\item{theme_options}{(Optional) list of arguments to pass to
vthemes::theme_vmodern().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_point()].}
}
\value{
A ggplot object
}
\description{
Creates a horizontal dot plot using custom ggplot theme, as an
  alternative to grouped bar charts.
}
\examples{
df <- data.frame(ID = rep(paste0("ID", 1:3), 2),
                 Group = rep(paste0("Group", LETTERS[1:2]), each = 3),
                 Value = c(1, 2, 3, -1, 0, 4))
plot_horizontal_dotplot(df, x_str = "Value", y_str = "ID", color_str = "Group")

}
